+++ apps/CMakeLists.txt
@@ -6,3 +6,4 @@ ADD_SUBDIRECTORY( UnitTwoDLib )
 ADD_SUBDIRECTORY( PerformanceGeom )
 ADD_SUBDIRECTORY( Projection )
 ADD_SUBDIRECTORY( MatrixGenerator )
+ADD_SUBDIRECTORY( TvbModels )
